Discounts
This class represents discounts available in the restaurant system. It handles CRUD.

createDiscount.php:
This endpoint is used to create a new discount. It accepts POST requests with JSON data containing details of the discount to be created. It uses the create() method of the Discounts class.

updateDiscount.php:
This endpoint is used to update an existing discount. It accepts PUT requests with JSON data containing the updated details of the discount. It uses the update() method of the Discounts class.

deleteDiscount.php:
 This endpoint is used to delete a discount. It accepts DELETE requests with JSON data containing the ID of the discount to be deleted. It uses the delete() method of the Discounts class.