working on booking from the client side
delete booking:
 It handles incoming requests, cancels the booking using the delete() method of the Booking class, and sends a JSON response indicating the status of the cancellation.

createbooking:
this method receives data representing the details of the booking, such as user ID, date, time, party size, preferences, and discount information. It then sanitizes the input data, prepares an SQL query to insert the data into the database, and executes the query.

update booking:
receives updated data representing the details of the booking, such as user ID, date, time, party size, preferences, and discount information. It then sanitises the input data, prepares an SQL query to update the corresponding record in the database, and executes the query.